feat(home): host filters/radius on home + fix sort-by-distance
Move the radius slider and filter section out of Settings → Search and
onto the home page below the saved-locations row. Drop the POI category
list. Settings → Search keeps only "Use device location" and default
location.

Fix sort by distance: pair RankPreference.DISTANCE with the required
setLocationRestriction(CircularBounds) instead of the old
setLocationBias(RectangularBounds), which silently produced empty
results. Use CircularBounds for all sort modes; treat the 20mi slider
max as a UI ceiling (use bias for relevance/rating, expand to 100mi
restriction for distance) so users aren't walled off from rare distant
matches.

Make filters live: SearchViewModel now invalidates its cached results
and re-runs the active query when any filter pref changes (radius,
open-now, open-in-1h, sort, max price). Display-only prefs
(showRating, distanceUnit) are excluded.

Drop dead Poi.isOutsideSearchRadius field and its UI block — never set.
