+++ app/src/main/java/com/example/helloworld/SearchScreen.kt
@@ -130,16 +130,6 @@ fun SearchScreen(
                         maxLines = 1,
                         overflow = androidx.compose.ui.text.style.TextOverflow.Ellipsis
                     )
-
-                    if (poi.isOutsideSearchRadius) {
-                        Spacer(modifier = Modifier.height(2.dp))
-                        Text(
-                            text = "Outside search radius",
-                            fontSize = 12.sp,
-                            color = MaterialTheme.colorScheme.error,
-                            fontWeight = FontWeight.SemiBold
-                        )
-                    }
                 }
                 Icon(
                     imageVector = Icons.AutoMirrored.Filled.ArrowForwardIos,
